Improve loggings for connections not coming from 127.0.0.1
Vserver environments and broken iptables rules can cause connections to come into the 'MailWatch SQL' process from an IP other than 127.0.0.1. The perl code silently drops those connections and so it is hard to debug.

Line added to MailWatch.pm to generate a maillog message for this event.

Thanks to Chris Bennett (cgb)