Add autocomplete="off" to the login form (#326)
Autocomplete enabled for sensitive HTML form fields

The Web form contains passwords or other sensitive text fields for which the browser auto-complete feature is enabled. Auto-complete stores completed form field and passwords locally in the browser, so that these fields are filled automatically when the user visits the site again.

Sensitive data and passwords can be stolen if the user's system is compromised.

Note, however, that form auto-complete is a non-standard, browser-side feature that each browser handles differently. Opera, for example, disregards the feature, requiring the user to enter credentials for each Web site visit.