Use domain admin username as domain filter
Domain administrators that have their domain name as user name are not able to set whitelist or blacklist entries for their domain unless it is explicitly added to the filters of the domain administrator user. 
However doing so results in duplicate email in the daily spam report.
This patch adds the domain administrator user name as domain filter.