improve quarantine_report
Having administrator users defined with one or more filter adresses (e.g. automatically imported from LDAP with mailwatch_ldap_sync.sh) led to quarantine reports with duplicate lines.

So when the user type is 'A' we must NOT iterate through the filters because administrators get a report for every single recipient anyway.

Also the administrator's report only had a partial headline because a suitable string was missing ($list_for was set empty). I suggest using the system's hostname in this case.

When sending _combined_ user reports the headline should only contain adresses which are contained in the currenct report. Otherwise this can get very messy when you have multiple adress filters defined for a user.