Enhancement for individual nonspam quarantine keeping (#1281)
I found it useful to have quarantine keeping days independently configured for nonspam.

So with this you can for example keep the spams for 30 days but the nonspams only for let's say 10 days.
On busy systems this can save performance and drive space.

The other way around is also possible if needed: e.g. keep spam 20 days and nonspam 40 days.